refactor: input 관련 패키지 분리
패키지 가독성 위해 Input 관련 패키지 생성 및 분리함.
