+++ src/main/java/racingcar/Main.java
@@ -1,7 +1,7 @@
 package racingcar;
 
-import racingcar.domain.InputCarCount;
-import racingcar.domain.InputRoundCount;
+import racingcar.domain.input.InputCarCount;
+import racingcar.domain.input.InputRoundCount;
 import racingcar.domain.RacingGame;
 import racingcar.view.InputView;
