+++ src/test/java/racingcar/domain/InputCarCountTest.java
@@ -2,6 +2,7 @@ package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.input.InputCarCount;
 
 import static org.assertj.core.api.Assertions.*;
