+++ src/test/java/racingcar/domain/InputRoundCountTest.java
@@ -2,7 +2,8 @@ package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.InputRoundCount;
+import racingcar.domain.input.InputCarCount;
+import racingcar.domain.input.InputRoundCount;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
