+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -2,6 +2,8 @@ package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.input.InputCarCount;
+import racingcar.domain.input.InputRoundCount;
 
 import static org.assertj.core.api.Assertions.*;
